piece obstacle checking works

+++ ChessEngine/ChessEngine/Pieces/Bishop.cs
@@ -20,9 +20,9 @@ namespace ChessEngine
         override
         public bool IsValidMove(Point p)
         {
-            int slope = 0;
-            int dx = (int)p.X - (int)this.Location.X;
-            int dy = (int)p.Y - (int)this.Location.Y;
+            double slope = 0;
+            double dx = p.X - this.Location.X;
+            double dy = p.Y - this.Location.Y;
 
             if (dx != 0) slope = dy / dx;
